Move Koin init to app launch
