+++ shared/src/androidMain/kotlin/org/jetbrains/kotlinconf/platformModule.android.kt
@@ -1,28 +1,18 @@
 package org.jetbrains.kotlinconf
 
 import android.app.Application
-import androidx.activity.ComponentActivity
-import androidx.activity.result.contract.ActivityResultContracts.RequestMultiplePermissions
 import androidx.preference.PreferenceManager
 import com.russhwolf.settings.ObservableSettings
 import com.russhwolf.settings.SharedPreferencesSettings
-import kotlinx.coroutines.channels.Channel
 import org.jetbrains.kotlinconf.utils.AndroidLogger
 import org.jetbrains.kotlinconf.utils.Logger
 import org.koin.core.module.Module
 import org.koin.dsl.module
 
 fun platformModule(
-    activity: ComponentActivity,
     application: Application,
     notificationIconId: Int,
 ): Module {
-    val permissionResult = Channel<Boolean>()
-    val permissionLauncher = activity.registerForActivityResult(RequestMultiplePermissions()) { result ->
-        val allGranted = result.all { it.value }
-        permissionResult.trySend(allGranted)
-    }
-
     return module {
         single<ObservableSettings> {
             SharedPreferencesSettings(PreferenceManager.getDefaultSharedPreferences(application))
@@ -31,8 +21,7 @@ fun platformModule(
             AndroidNotificationService(
                 timeProvider = get(),
                 context = application,
-                permissionResult = permissionResult,
-                permissionLauncher = permissionLauncher,
+                permissionHandler = get(),
                 iconId = notificationIconId,
                 logger = get(),
             )
