+++ shared/src/wasmJsMain/kotlin/org/jetbrains/kotlinconf/main.wasm.kt
@@ -2,12 +2,7 @@ package org.jetbrains.kotlinconf
 
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.window.CanvasBasedWindow
-import com.russhwolf.settings.ExperimentalSettingsApi
-import com.russhwolf.settings.ObservableSettings
-import com.russhwolf.settings.StorageSettings
-import com.russhwolf.settings.observable.makeObservable
 import org.jetbrains.kotlinconf.ui.initCoil
-import org.koin.dsl.module
 
 @JsModule("@js-joda/timezone")
 external object JsJodaTimeZoneModule
@@ -17,13 +12,9 @@ private val jsJodaTz = JsJodaTimeZoneModule
 @OptIn(ExperimentalComposeUiApi::class)
 fun main() {
     initCoil()
+    initKoin(platformModule)
+
     CanvasBasedWindow {
-        App(platformModule = module {
-            single<ObservableSettings> {
-                @OptIn(ExperimentalSettingsApi::class)
-                StorageSettings().makeObservable()
-            }
-            single<NotificationService> { ServiceWorkerNotificationService(get()) }
-        })
+        App()
     }
 }
